+++ eeprom.h
@@ -3,7 +3,7 @@
 #define uchar unsigned char
 #define uint unsigned int
 void EEPROM_write(uint addr_h,uint addr_l,uint data);
-uchar EEPROM_read(uint addr_h,uint addr_l);
+uint EEPROM_read(uint addr_h,uint addr_l);
 
 
 #endif
